+++ apps/core/context_processors.py
@@ -1,2 +1,8 @@
+from django.conf import settings
+
+
 def application_meta(_request):
-    return {"application_name": "BrickMissing", "application_version": "8.0.0"}
+    return {
+        "application_name": "BrickMissing",
+        "application_version": settings.BRICKMISSING_VERSION,
+    }